applied suggested changes
