+++ unity-renderer/Assets/Scripts/MainScripts/DCL/Controllers/HUD/NotificationMessagesHUD/NotificationComponent/IChatNotificationMessageComponentView.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 91019bb8140be41a2b2c12096164d4e3
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
